temp/requirements.txt
matplotlib==3.1.2
numpy==1.17.4
Pillow==9.2.0
tensorflow==2.9.1
tensorflow_datasets==4.6.0
torch==1.12.0
torchvision==0.13.0
tqdm==4.64.0
vit_pytorch==0.35.8
